Refactored to add BaseAutonomous
